refactor: extract type SecretNameToSNIs
